+++ Cargo.toml
@@ -90,6 +90,7 @@ members = [
     "nexus/authz-macros",
     "nexus/auth",
     "nexus/background-task-interface",
+    "nexus/bootstore-config",
     "nexus/db-errors",
     "nexus/db-fixed-data",
     "nexus/db-lookup",
@@ -280,6 +281,7 @@ default-members = [
     "nexus/authz-macros",
     "nexus/auth",
     "nexus/background-task-interface",
+    "nexus/bootstore-config",
     "nexus/db-errors",
     "nexus/db-fixed-data",
     "nexus/db-lookup",
@@ -622,6 +624,7 @@ ddm-admin-client = { git = "https://github.com/oxidecomputer/maghemite", rev = "
 multimap = "0.10.1"
 nexus-auth = { path = "nexus/auth" }
 nexus-background-task-interface = { path = "nexus/background-task-interface" }
+nexus-bootstore-config = { path = "nexus/bootstore-config" }
 nexus-client = { path = "clients/nexus-client" }
 nexus-config = { path = "nexus-config" }
 nexus-db-errors = { path = "nexus/db-errors" }
