+++ nexus/bootstore-config/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "nexus-bootstore-config"
+version = "0.1.0"
+edition.workspace = true
+
+[lints]
+workspace = true
+
+[build-dependencies]
+omicron-rpaths.workspace = true
+
+[dependencies]
+ipnetwork.workspace = true
+nexus-db-model.workspace = true
+nexus-db-queries.workspace = true
+omicron-common.workspace = true
+oxnet.workspace = true
+# See omicron-rpaths for more about the "pq-sys" dependency.
+pq-sys = "*"
+sled-agent-types.workspace = true
+slog.workspace = true
+slog-error-chain.workspace = true
+uuid.workspace = true
+
+omicron-workspace-hack.workspace = true
